+++ src/values/user/registrationFormValues.ts
@@ -1,4 +1,4 @@
-import type { UserRegistrationFormFields } from '../../schema/user/registrationFormSchema';
+import type { UserRegistrationFormFields } from '../../schemas/user/registrationFormSchema';
 
 export const initialValues: UserRegistrationFormFields = {
   firstName: '',
